Crate rsynth

Rsynth

The rsynth crate makes it easier to write software synthesizers.

The Plugin trait

The functionality of the plugin that is common to all back-ends is defined by the Plugin trait. It defines some meta-data about the plugin and contains the methods for rendering the audio.

The EventHandler trait

Plugins also implement EventHandler for each event type that they support. Currently supported events are:

RawMidiEvent

SysExEvent

Back-ends

rsynth currently supports two back-ends:

jack

vst

In order to support a specific back-end, plugins may additionally need to implement a backend-specific trait on top of the Plugin trait. See the documentation of each back-end for more information.

Middleware

You can add features, such as polyphony, to your plug-in by using middleware. Typically, suppose M is middleware and your plugin P implement the Plugin trait and any other backend-specific trait, then M<P> also implements the Plugin trait and the backend-specific traits P implements. Similar for the EventHandler trait: when your plugin P implements EventHandler<E>, then M<P> typically does the same.

Currently, supported middleware is

Polyphony

ZeroInit

Context

Context can be used to access data that is not stored in your plugin, but e.g. stored by the host or some middleware. The context is passed to every call to render_buffer in the Plugin trait and to every call to handle_event in the EventHandler trait.

Modules

backend
Defines the different backends.

context
Traits and macros for working with context.

dev_utilities
Utilities to be used when developing backends and middleware.

envelope

event
This module defines the EventHandler trait and some event types: RawMidiEvent, SysExEvent, ...

middleware
Middleware like support for polyphony, ...

note

output_mode
Utilities to handle both polyphonic and monophonic plugins.

Macros

macro_for_application

macro_for_rsynth

macro_for_rsynth_feature_jack

macro_for_rsynth_feature_vst

macro_for_unpublished_crate

vst_init
A wrapper around the plugin_main! macro from the vst crate. You call this with one parameter, which is the function declaration of a function that creates your plugin. This function may also do some setup (e.g. initialize logging).

wrap_context
wrap_context!($wrapper_type, $type_name, $(,$type_param)*); generates a type with the name $wrapper_type that adds an already-existing type $type_name to a "child_context".

Traits

Plugin
The trait that all plugins need to implement. The parameter C corresponds to the context of the plugin.