Module rsynth::backend

Backends.

Pre-defined backends

rsynth currently supports the following back-ends:

combined combine different back-ends for audio input, audio output, midi input and midi output, mostly for offline rendering and testing (behind various features)

jack (behind the backend-jack feature)

vst (behind the backend-vst feature)

These backends are currently in the rsynth crate, but we may eventually move them to separate crates.

Custom backends

You can write a backend in a separate crate. If you encounter problems that prevent you from writing your backend in a separate crate (e.g., we have forgotten to mark something as pub), let us know by opening an issue.

Publishing a backend crate

When you publish a backend crate, let us know by opening an issue or pull request so that we can link to it in the documentation of rsynth.

Modules

combined
Combine different back-ends for audio input, audio output and midi input, mostly for offline rendering and testing.

jack_backend
Wrapper for the JACK backend (behind the backend-jack feature).

vst_backend
Wrapper for the VST backend (behind the backend-vst feature).

Traits

HostInterface
Defines an interface for communicating with the host or server of the backend, e.g. the VST host when using VST or the Jack server when using Jack.

Stop
A marker trait that indicates that the backend can be stopped.