Module rsynth::backend::combined

Combine different back-ends for audio input, audio output, midi input and midi output, mostly for offline rendering and testing.

Support is only enabled if rsynth is compiled with the "backend-combined" feature, see the cargo reference for more information on setting cargo features.

The run function can be used to run a plugin and read audio and midi from the inputs and write audio and midi to the outputs.

Currently, the following inputs and outputs are available:

Dummy: AudioDummy: dummy audio input (generates silence) and output and MidiDummy: dummy midi input (generates no events) and output

Hound: HoundAudioReader and HoundAudioWriter: read and write .wav files (behind the "backend-combined-hound" feature)

Rimd: RimdMidiReader and RimdMidiWriter: reand and write .mid files (behind the "backend-combined-rimd" feature, only available when using rsynth from sources, not from crates.io)

Memory: AudioBufferReader and AudioBufferWriter: read and write audio from memory

Testing: TestAudioReader and TestAudioWriter: audio input and output, to be used in tests

Modules

dummy
Dummy backend that does nothing, useful for testing.

hound
Backend for reading and writing .wav files, based on the hound crate.

memory
In-memory backend, useful for testing.

Structs

MidiWriterWrapper
A wrapper around a midi writer that implements EventHandler<Timed<RawMidiEvent>> by queueing the events, which can then be written to the encapsulated MidiWriter by calling step_frames.

TestAudioReader
An audio reader, useful for testing.

TestAudioWriter
An audio writer, useful for testing.

TestMidiReader
A midi reader, useful for testing.

TestMidiWriter
A midi writer, useful for testing.

Enums

CombinedError
The error type that represents the errors you can get from the run function.

Constants

MICROSECONDS_PER_SECOND

Traits

AudioReader
Define how audio is read.

AudioWriter
Define how audio is written.

MidiWriter

Functions

run
Run an audio renderer with the given audio input, audio output, midi input and midi output.