Module rsynth::backend::combined::memory::dasp_sample

Modules

conv
Pure functions and traits for converting between i8, i16, I24, i32, I48, i64, u8, u16, U24, u32, U48, u64, f32 and f64.

types
A collection of custom, non-std Sample types.

Structs

I24

I48

U24

U48

Traits

Duplex
Sample types which may be converted to and from some type S.

FloatSample
Sample format types represented as floating point numbers.

FromSample
Similar to the std From trait, but specifically for converting between sample types.

Sample
A trait for working generically across different Sample format types.

SignedSample
Integral and floating-point Sample format types whose equilibrium is at 0.

ToSample
Similar to the std Into trait, but specifically for converting between sample types.