Trait rsynth::backend::combined::memory::dasp_sample::FloatSample

pub trait FloatSample: Duplex<f32> + Duplex<f64> + Sample<Signed = Self, Float = Self> + SignedSample + Mul<Self, Output = Self> + Div<Self, Output = Self> {
    pub const IDENTITY: Self;

    pub fn sample_sqrt(self) -> Self;
}

Sample format types represented as floating point numbers.

Samples often need to be converted to some mutual FloatSample type for signal scaling and modulation.

Associated Constants

pub const IDENTITY: Self

Represents the multiplicative identity of the floating point signal.

Loading content...
Required methods

pub fn sample_sqrt(self) -> Self

Calculate the square root of Self.

Loading content...