Trait rsynth::backend::combined::memory::dasp_sample::FromSample

pub trait FromSample<S> {
    pub fn from_sample_(s: S) -> Self;
}

Similar to the std From trait, but specifically for converting between sample types.

We use this trait to be generic over the Sample::to_sample and Sample::from_sample methods.

Required methods

pub fn from_sample_(s: S) -> Self

Loading content...