Module rsynth::backend::combined::memory::wav_0_6

Modules

bit_depth
Contains items for dealing with the "data" chunk of wave files.

header
Contains items responsible for dealing with the "fmt " chunk of wave files.

Structs

Header
Structure for the "fmt " chunk of wave files, specifying key information about the enclosed data.

Enums

BitDepth
Enum listing the supported bit-depths and containers for the samples at each depth.

Constants

WAV_FORMAT_IEEE_FLOAT
Value signifying IEEE float data.

WAV_FORMAT_PCM
Value signifying PCM data.

Functions

read
Reads in the given reader and attempts to extract the audio data and header from it.

write
Writes the given wav data to the given writer.