Function rsynth::backend::combined::memory::wav_0_6::write

pub fn write<W>(
    header: Header, 
    track: &BitDepth, 
    writer: &mut W
) -> Result<(), Error> where
    W: Write + Seek, 

Writes the given wav data to the given writer.

Notes

Although track is a borrowed value, its contents will be formatted into an owned Vec<u8> so that it can be written to the writer through [riff::ChunkContents::write].

Errors

This function fails under the following circumstances:

Any error occurring from the writer parameter during writing.

The given BitDepth is BitDepth::Empty.