Module rsynth::backend::combined::midly::midly

Re-exports from the midly crate.

Modules

number
Special-length integers used by the MIDI standard.

Structs

Error
Represents an error parsing an SMF file or MIDI stream.

Event
Represents a fully parsed track event, with delta time.

Header
A MIDI file header.

Smf
Represents a Standard Midi File (.mid and .midi files). Yields TrackRepr on a Vec, allowing for customization on what is stored in a track. References an outside byte array.

SmpteTime
Encodes an SMPTE time of the day.

TrackIter
Allows deferring track parsing for later, on a per-event basis.

Enums

ErrorKind
The type of error that occurred while parsing.

EventKind
Represents the different kinds of events.

Format
The different formats an SMF file can be.

Fps

MetaMessage

MidiMessage
Represents a MIDI message, not an event.

Timing
The timing for an SMF file. This can be in ticks/beat or ticks/second.

Traits

TrackRepr
Allows for customization on how tracks are stored in memory.