Enum rsynth::backend::combined::midly::midly::MidiMessage

pub enum MidiMessage {
    NoteOff {
        key: u7,
        vel: u7,
    },
    NoteOn {
        key: u7,
        vel: u7,
    },
    Aftertouch {
        key: u7,
        vel: u7,
    },
    Controller {
        controller: u7,
        value: u7,
    },
    ProgramChange {
        program: u7,
    },
    ChannelAftertouch {
        vel: u7,
    },
    PitchBend {
        bend: PitchBend,
    },
}

Represents a MIDI message, usually associated to a MIDI channel.

If you wish to parse a MIDI message from a slice of raw MIDI bytes, use the LiveEvent::parse method instead and ignore all variants except for LiveEvent::Midi.

Variants

NoteOff

Stop playing a note.

Fields of NoteOff

key: u7

The MIDI key to stop playing.

vel: u7

The velocity with which to stop playing it.

NoteOn

Start playing a note.

Fields of NoteOn

key: u7

The key to start playing.

vel: u7

The velocity (strength) with which to press it.

Note that by convention a NoteOn message with a velocity of 0 is equivalent to a NoteOff.

Aftertouch

Modify the velocity of a note after it has been played.

Fields of Aftertouch

key: u7

The key for which to modify its velocity.

vel: u7

The new velocity for the key.

Controller

Modify the value of a MIDI controller.

Fields of Controller

controller: u7

The controller to modify.

See the MIDI spec for the meaning of each index.

value: u7

The value to set it to.

ProgramChange

Change the program (also known as instrument) for a channel.

Fields of ProgramChange

program: u7

The new program (instrument) to use for the channel.

ChannelAftertouch

Change the note velocity of a whole channel at once, without starting new notes.

Fields of ChannelAftertouch

vel: u7

The new velocity for all notes currently playing in the channel.

PitchBend

Set the pitch bend value for the entire channel.

Fields of PitchBend

bend: PitchBend

The new pitch-bend value.