Struct rsynth::backend::combined::midly::midly::io::Cursor

pub struct Cursor<'a> { /* fields omitted */ }

A seekable writer over an in-memory buffer.

Available even when the std and alloc features are disabled.

Implementations

impl<'a> Cursor<'a>

pub fn new(buffer: &mut [u8]) -> Cursor<'_>

Create a new cursor located at the start of the given buffer.

pub fn from_parts(buffer: &mut [u8], cursor: usize) -> Cursor<'_>

Create a cursor from a buffer and the cursor within it.

Panics

Panics if cursor > buffer.len().

pub fn into_parts(self) -> (&'a mut [u8], usize)

Yield the underlying buffer and the cursor within it.

The cursor is guaranteed to be cursor <= buffer.len().

pub fn slice(&self) -> &[u8]

Get a reference to the whole underlying buffer.

pub fn slice_mut(&mut self) -> &mut [u8]

Get a mutable reference to the whole underlying buffer.

pub fn cursor(&self) -> usize

Get the position of the cursor.

pub fn written(&self) -> &[u8]

Get a reference to the written portion of the buffer.

pub fn unwritten(&self) -> &[u8]

Get a reference to the portion of the buffer that is not yet written.

pub fn split(&self) -> (&[u8], &[u8])

Split the buffer into the written and unwritten parts.

pub fn written_mut(&mut self) -> &mut [u8]

Get a mutable reference to the written portion of the buffer.

pub fn unwritten_mut(&mut self) -> &mut [u8]

Get a mutable reference to the portion of the buffer that is not yet written.

pub fn split_mut(&mut self) -> (&mut [u8], &mut [u8])

Split the buffer into the written and unwritten parts.