Trait rsynth::backend::combined::midly::midly::io::Seek

pub trait Seek: Write {
    pub fn tell(&mut self) -> Result<u64, Self::Error>;

    pub fn write_at(&mut self, buf: &[u8], pos: u64) -> Result<(), Self::Error>;
}

A Seek trait available even in no_std environments.

Not all of the Seek functionality is required, only the functionality required to write MIDI files.

Required methods

pub fn tell(&mut self) -> Result<u64, Self::Error>

Where is the writer currently at.

pub fn write_at(&mut self, buf: &[u8], pos: u64) -> Result<(), Self::Error>

Write a slice of data at the given absolute position, and return to the end of the writer afterwards.

Loading content...