Module rsynth::backend::combined::midly::midly::stream

Provides support for the niche use case of reading MIDI events from a non-delimited stream.

Low level drivers or embedded devices might have access to raw MIDI streams, which are just a stream of bytes that might arrive at any rate.

For this use case, the MidiStream type is provided, which can receive streaming bytes and produces MIDI messages.

Handles all of the quirks specific to MIDI streams, including System Realtime messages embedded in the middle of another message.

Structs

DefaultBuffer
The default buffer type used for MidiStream. By default it will have a reasonable maximum capacity, but the Buffer trait can be implemented for fine-grained control.

MidiStream
A streaming raw MIDI parser. This parser takes raw MIDI, not .midi files!

Traits

Buffer
Describes types that can be used as data buffers for the MidiStream type.