Module rsynth::backend::combined::midly::midly_0_5::io

Provides abstractions over writers, even in no_std environments.

When the std feature is enabled, IoWrap and SeekableWrap provide a bridge between the std::io API and the midly::io API. Besides, write methods that work with midly::io::Write types usually provide a write_std variant that works with std::io::Write types when the std feature is enabled.

Structs

Cursor
A seekable writer over an in-memory buffer.

IoWrap
Bridge between a midly::io::Write type and a std::io::Write type.

NotSeekable
The type used for the Seekable associated type on non-seekable writers.

SeekableWrap
Bridge between a midly::io::{Write, Seek} type and a std::io::{Write, Seek} type.

Enums

CursorError
The errors that can arise when writing to an in-memory buffer.

Traits

Seek
A Seek trait available even in no_std environments.

Write
A Write trait available even in no_std environments, and with per-type errors.

Type Definitions

WriteResult
Either Ok(()) or the error specific to the W writer.