Trait rsynth::backend::combined::AudioReader

pub trait AudioReader<F> {
    fn number_of_channels(&self) -> usize;

    fn frames_per_second(&self) -> u64;

    fn fill_buffer(&mut self, output: &mut [&mut [F]]) -> usize;
}

Required methods

fn number_of_channels(&self) -> usize

fn frames_per_second(&self) -> u64

fn fill_buffer(&mut self, output: &mut [&mut [F]]) -> usize

Fill the buffers. Return the number of frames that have been written. If it is < the number of frames in the input, now more frames can be expected.

Loading content...