Trait rsynth::backend::combined::AudioWriter

pub trait AudioWriter<S> where
    S: Copy, {
    type Err;
    fn write_buffer(
        &mut self, 
        buffer: &AudioBufferIn<S>
    ) -> Result<(), Self::Err>;
}

Define how audio is written.

This trait is generic over S, which represents the data-type used for a sample.

Associated Types

type Err

The type of the error that occurs when reading data.

Loading content...
Required methods

fn write_buffer(&mut self, buffer: &AudioBufferIn<S>) -> Result<(), Self::Err>

Loading content...