Function rsynth::backend::jack_backend::run

pub fn run<P>(plugin: P) -> Result<P, Error> where
    P: CommonPluginMeta + AudioHandler + CommonAudioPortMeta + CommonMidiPortMeta + Send + Sync + 'static,
    P: ContextualAudioRenderer<f32, JackHost<'c, 'mp, 'mw>>,
    P: ContextualEventHandler<Indexed<Timed<RawMidiEvent>>, JackHost<'c, 'mp, 'mw>>,
    P: ContextualEventHandler<Indexed<Timed<SysExEvent<'a>>>, JackHost<'c, 'mp, 'mw>>, 

Run the plugin until the user presses a key on the computer keyboard or the plugin requests the JackHost to stop.