Struct rsynth::backend::jack_backend::jack::ProcessScope

pub struct ProcessScope { /* fields omitted */ }

ProcessScope provides information on the client and frame time information within a process callback.

Implementations

impl ProcessScope

pub fn n_frames(&self) -> u32

The number of frames in the current process cycle.

pub fn last_frame_time(&self) -> u32

The precise time at the start of the current process cycle. This function may only be used from the process callback, and can be used to interpret timestamps generated by self.frame_time() in other threads, with respect to the current process cycle.

pub fn frames_since_cycle_start(&self) -> u32

The estimated time in frames that has passed since the JACK server began the current process cycle.

pub fn cycle_times(&self) -> Result<CycleTimes, Error>

Provides the internal cycle timing information as used by most of the other time related functions. This allows the caller to map between frame counts and microseconds with full precision (i.e. without rounding frame times to integers), and also provides e.g. the microseconds time of the start of the current cycle directly (it has to be computed otherwise).

Err(Error::TimeError) is returned on failure. Err(Error::WeakFunctionNotFound) if the function does not exist.

pub fn client_ptr(&self) -> *mut Struct__jack_client

Expose the client_ptr for low level purposes.

This is mostly for use within the jack crate itself.

pub unsafe fn from_raw(
n_frames: u32,
client_ptr: *mut Struct__jack_client
) -> ProcessScope

Create a ProcessScope for the client with the given pointer and the specified amount of frames.

This is mostly for use within the jack crate itself.