Struct rsynth::backend::jack_backend::jack::RingBufferWriter

pub struct RingBufferWriter { /* fields omitted */ }

Write end of the ring buffer. Can only be used from one thread (can be a different from the read thread).

Implementations

impl RingBufferWriter

pub fn write_buffer(&mut self, buf: &[u8]) -> usize

Write data into the ringbuffer. Returns: The number of bytes written, which may range from 0 to buf.len()

pub fn advance(&mut self, cnt: usize)

Advance the write pointer. use this after peek_iter or get_vector to advance the buffer pointer.

pub fn space(&mut self) -> usize

Return the number of bytes available for writing.

pub fn get_vector(&mut self) -> (&mut [u8], &mut [u8])

Return a pair of slices of the current writable space in the ringbuffer. two slices are needed because the space available for writing may be split across the end of the ringbuffer. consider using peek_iter for convenience.

pub fn peek_iter(&'a mut self) -> Chain<IterMut<'a, u8>, IterMut<'a, u8>>

Iterator that goes over all the data available to write.