Struct rsynth::backend::jack_backend::jack::TransportBBT

pub struct TransportBBT {
    pub bar: usize,
    pub beat: usize,
    pub tick: usize,
    pub sig_num: f32,
    pub sig_denom: f32,
    pub ticks_per_beat: f64,
    pub bpm: f64,
    pub bar_start_tick: f64,
}

Transport Bar Beat Tick data.

Fields
bar: usize

Time signature bar, 1 or more.

beat: usize

Time signature beat, 1 <= beat <= sig_num.

tick: usize

current tick-within-beat

Remarks

Should be >= 0 and < ticks_per_beat: the first tick is tick 0.

sig_num: f32

Time Signature "numerator". Jack calls this beats_per_bar.

sig_denom: f32

Time Signature "denominator". Jack calls this beat_type.

ticks_per_beat: f64

Number of ticks within a beat.

Remarks

Usually a moderately large integer with many denominators, such as 1920.0

bpm: f64

BPM, quantized to block size. This means when the tempo is not constant within this block, the BPM value should adapted to compensate for this. This is different from most fields in this struct, which specify the value at the beginning of the block rather than an average.

bar_start_tick: f64

Number of ticks that have elapsed between frame 0 and the first beat of the current measure.

Implementations

impl TransportBBT

pub fn with_bbt(
&mut self,
bar: usize,
beat: usize,
tick: usize
) -> &mut TransportBBT

Set bar, beat, tick

Example

use jack::TransportBBT;
let bbt = TransportBBT::default().with_bbt(4, 2, 14).validated();
assert!(bbt.is_ok());
let bbt = bbt.unwrap();
assert_eq!(bbt.bar, 4);
assert_eq!(bbt.beat, 2);
assert_eq!(bbt.tick, 14);

pub fn with_bpm(&mut self, bpm: f64) -> &mut TransportBBT

Set Beats Per Minute.

pub fn with_timesig(&mut self, num: f32, denom: f32) -> &mut TransportBBT

Set the time signature.

pub fn with_ticks_per_beat(&mut self, ticks_per_beat: f64) -> &mut TransportBBT

Set ticks per beat.

pub fn with_bar_start_tick(&mut self, tick: f64) -> &mut TransportBBT

Set bar start tick.

pub fn validated(&self) -> Result<TransportBBT, TransportBBTValidationError>

Validate contents.

pub fn valid(&self) -> bool