Trait rsynth::backend::jack_backend::jack::PortSpec

pub unsafe trait PortSpec {
    fn jack_port_type(&self) -> &str;

    fn jack_flags(&self) -> PortFlags;

    fn jack_buffer_size(&self) -> u64;
}

Defines the configuration for a certain port to JACK, ie 32 bit floating audio input, 8 bit raw midi output, etc...

Required methods

fn jack_port_type(&self) -> &str

String used by JACK upon port creation to identify the port type.

fn jack_flags(&self) -> PortFlags

Flags used by jack upon port creation.

fn jack_buffer_size(&self) -> u64

Size used by jack upon port creation.

Loading content...