Module rsynth::backend::vst_backend

Wrapper for the VST backend (behind the backend-vst feature).

Support is only enabled if you compile with the "backend-vst" feature, see the cargo reference for more information on setting cargo features.

For an example, see vst_synth.rs in the examples folder. examples/example_synth contains the code that is shared for all backends and examples/vst_synth.rs contains the jack-specific code.

Usage

See the documentation of the vst_init macro.

Modules

vst
Re-exports from the vst-rs crate. Use this in libraries so that your library does not break when rsynth upgrades to another version of the vst-rs crate.

Structs

VstPluginWrapper
A struct used internally by the vst_init macro. Normally, plugin's do not need to use this.

Traits

VstPluginMeta
Define some VST-specific meta-data for a VST plugin.