Module rsynth::backend::vst_backend::vst

Re-exports from the vst-rs crate. Use this in libraries so that your library does not break when rsynth upgrades to another version of the vst-rs crate.

Modules

api
Structures and types for interfacing with the VST 2.4 API.

buffer
Buffers to safely work with audio samples.

channels
Meta data for dealing with input / output channels. Not all hosts use this so it is not necessary for plugin functionality.

editor
All VST plugin editor related functionality.

event
Interfaces to VST events.

host
Host specific structures.

plugin
Plugin specific structures.

util
Structures for easing the implementation of VST plugins.

Macros

plugin_main
Exports the necessary symbols for the plugin to be used by a VST host.