Struct rsynth::buffer::AudioChunk

pub struct AudioChunk<S> { /* fields omitted */ }

Implementations

impl<S> AudioChunk<S>

pub fn zero(number_of_channels: usize, number_of_frames: usize) -> Self where
S: Zero,

Note: cannot be used in a real-time context

This method allocates memory and cannot be used in a real-time context.

pub fn from_channels(channels: Vec<Vec<S>>) -> Self

pub fn new(number_of_channels: usize) -> Self

Note: cannot be used in a real-time context

This method allocates memory and cannot be used in a real-time context.

pub fn with_capacity(number_of_channels: usize, capacity: usize) -> Self

Create a new AudioChunk in which each channel has the given capacity. This allows to append capacity frames to the AudioChunk (e.g. by calling append_sliced_chunk).

Note: cannot be used in a real-time context

This method allocates memory and cannot be used in a real-time context.

pub fn channels(&self) -> &Vec<Vec<S>>

pub fn append_sliced_chunk(&mut self, chunk: &[&[S]]) where
S: Clone,

Note about using in a real-time context

This method will allocate memory if the capacity of the chunk is exceeded and cannot be used in a real-time context in that case.

pub fn inner(self) -> Vec<Vec<S>>

pub fn as_slices<'a>(&'a self) -> Vec<&[S]>

Note: cannot be used in a real-time context

This method allocates memory and cannot be used in a real-time context.

pub fn as_mut_slices<'a>(&'a mut self) -> Vec<&mut [S]>

Note: cannot be used in a real-time context

This method allocates memory and cannot be used in a real-time context.

pub fn split(self, number_of_frames_per_chunk: usize) -> Vec<Self>

Note: cannot be used in a real-time context

This method allocates memory and cannot be used in a real-time context.