Module rsynth::event

Event handling

This module defines the EventHandler trait and some event types: RawMidiEvent, SysExEvent, ...

Custom events

Implement Copy if possible

If possible, implement the Copy trait for the event, so that the event can be dispatched to different voices in a polyphonic context.

Modules

event_queue
Queue events.

Structs

DeltaEvent

Indexed
Indexed<E> adds an index to an event of type E. The index typically corresponds to the index of the channel.

Map
An EventHandler from the EventHandlerExt::map method.

RawMidiEvent
A raw midi event.

SysExEvent
A System Exclusive ("SysEx") event.

TimeStretcher
Stretch integer (u64) time stamps by a fractional factor that may change over time.

Timed
Timed<E> adds timing to an event.

Enums

MidlyConversionError
The error type when converting from midly's TrackEventKind to a RawMidiEvent.

Traits

ContextualEventHandler
The trait that plugins should implement in order to handle the given type of events.

EventHandler
The trait that plugins should implement in order to handle the given type of events.

EventHandlerExt
An extension trait for EventHandler providing some convenient combinator functions.