Trait rsynth::event::ContextualEventHandler

pub trait ContextualEventHandler<E, Context> {
    fn handle_event(&mut self, event: E, context: &mut Context);
}

The trait that plugins should implement in order to handle the given type of events.

The type parameter E corresponds to the type of the event. The type parameter Context refers to the context that is passed to the event handler.

Required methods

fn handle_event(&mut self, event: E, context: &mut Context)

Loading content...