Trait rsynth::event::EventHandler

pub trait EventHandler<E, C> {
    fn handle_event(&mut self, event: E, context: &mut C);
}

The trait that plugins should implement in order to handle the given type of events.

The type parameter E corresponds to the type of the event. The type parameter C corresponds to the context of the plugin.

Required methods

fn handle_event(&mut self, event: E, context: &mut C)

Loading content...