Macro rsynth::vst_init

macro_rules! vst_init {
    (fn $function_name:ident() -> $return_type:ty
        $body:block
    ) => { ... };
}

A wrapper around the plugin_main! macro from the vst crate. You call this with one parameter, which is the function declaration of a function that creates your plugin. This function may also do some setup (e.g. initialize logging).

Example:

struct MyPlugin {
  // Define your fields here
}

use rsynth::{
    CommonAudioPortMeta,
    CommonPluginMeta,
    event::{
        ContextualEventHandler,
        Timed,
        RawMidiEvent,
        SysExEvent
    },
    backend::{
        HostInterface,
        vst_backend::VstPluginMeta
    },
    ContextualAudioRenderer,
    AudioHandlerMeta,
    AudioHandler
};
use vst::plugin::Category;
impl VstPluginMeta for MyPlugin {
    // Implementation omitted for brevity.
}

use asprim::AsPrim;
use num_traits::Float;

impl AudioHandlerMeta for MyPlugin {
     // Implementation omitted for brevity.
}

impl AudioHandler for MyPlugin {
    // Implementation omitted for brevity.
}

impl CommonPluginMeta for MyPlugin {
    // Implementation omitted for brevity.
}
impl CommonAudioPortMeta for MyPlugin
{
    // Implementation omitted for brevity.
}

impl<F, H> ContextualAudioRenderer<F, H> for MyPlugin
where
    F: Float + AsPrim,
    H: HostInterface,
{
    // Implementation omitted for brevity.
}

impl<H> ContextualEventHandler<Timed<RawMidiEvent>, H> for MyPlugin
where
    H: HostInterface,
{
    // Implementation omitted for brevity.
}
impl<'a, H> ContextualEventHandler<Timed<SysExEvent<'a>>, H> for MyPlugin
where
    H: HostInterface,
{
    // Implementation omitted for brevity.
}
vst_init!(
   fn init() -> MyPlugin {
       MyPlugin{}
   }
);