Macro rsynth::wrap_context

macro_rules! wrap_context {
    ($wrapper_name:ident, $type_name:ty $(,$type_param:ident)*) => { ... };
}

wrap_context!($wrapper_type, $type_name, $(,$type_param)*); generates a type with the name $wrapper_type that adds an already-existing type $type_name to a "child_context".

The generated type looks as follows:

pub struct $wrapper_name<'a, 'c, C $(,$type_param)*> {
     aspect: &'a mut $type_name,
     child_context: &'c mut C,
 }

The generated type implements TransparentContext<$type_name> by returning the field aspect and TransparentContext<T> for "any other" type T for which the child context implements Transparentcontext<T> by delegating it to the child context.

The generated type has the following two fields

aspect, which is a reference as mutable to the given type, and

child_context, which is a reference as mutable to the "child context".

The generated type has also a non-public new function as follows:

fn new(aspect: &'a mut $type_name, child_context: &'c mut C) -> Self;

Examples

struct MyContextAspect {
    // ...
}

wrap_context!(MyContext, MyContextAspect);

If the type has type parameters, these must be added at the end:

struct MyContextAspect<T1, T2> {
    // ...
}

wrap_context!(MyContext, MyContextAspect<T1, T2>, T1, T2);