Trait rsynth::meta::Name

pub trait Name {
    fn name(&self) -> &str { ... }

    fn write_name<W: Write>(&self, buffer: &mut W) -> Result<(), Error> { ... }
}

Implement this trait to indicate that a type can be used to represent meta-data that contains a name.

Provided methods

fn name(&self) -> &str

👎 Deprecated since 0.1.2:
Use or implement write_name instead.

Get the name.

fn write_name<W: Write>(&self, buffer: &mut W) -> Result<(), Error>

Write the name to the given buffer.

Compatibility note

The default implementation of this method will likely be removed in a future release.

Loading content...