Trait rsynth::middleware::polyphony::VoiceStealMode

pub trait VoiceStealMode {
    type V;
    type State: Sized + Default;
    fn find_idle_voice<'v>(
        &mut self, 
        voices: &'v mut [VoiceWithState<Self::V, Self::State>], 
        note: u8
    ) -> &'v mut VoiceWithState<Self::V, Self::State>;

    fn find_voice_playing_note<'v>(
        &mut self, 
        voices: &'v mut [VoiceWithState<Self::V, Self::State>], 
        note: u8
    ) -> Option<&'v mut VoiceWithState<Self::V, Self::State>>;

    fn mark_voice_as_active(
        &mut self, 
        voice: &mut VoiceWithState<Self::V, Self::State>, 
        note: u8
    );

    fn mark_voice_as_inactive(
        &mut self, 
        voice: &mut VoiceWithState<Self::V, Self::State>
    );
}

Implement this trait to define your own VoiceStealMode.

Associated Types

type V

The type of the voice (implementing the Voice trait) that this VoiceStealMode can handle.

type State: Sized + Default

A data type to store the state of the voice (idle, ...) if needed.

Loading content...
Required methods

fn find_idle_voice<'v>(
&mut self,
voices: &'v mut [VoiceWithState<Self::V, Self::State>],
note: u8
) -> &'v mut VoiceWithState<Self::V, Self::State>

Decide which voice should handle a given note. This method is at the heart of the VoiceStealMode.

fn find_voice_playing_note<'v>(
&mut self,
voices: &'v mut [VoiceWithState<Self::V, Self::State>],
note: u8
) -> Option<&'v mut VoiceWithState<Self::V, Self::State>>

Return the voice that is playing a given note, if any.

fn mark_voice_as_active(
&mut self,
voice: &mut VoiceWithState<Self::V, Self::State>,
note: u8
)

Mark this voice as "active".

fn mark_voice_as_inactive(
&mut self,
voice: &mut VoiceWithState<Self::V, Self::State>
)

Mark the given voice as "inactive".

Loading content...