Trait rsynth::AudioHandlerMeta

pub trait AudioHandlerMeta {
    fn max_number_of_audio_inputs(&self) -> usize;

    fn max_number_of_audio_outputs(&self) -> usize;
}

Define the maximum number of audio inputs and the maximum number of audio outputs.

Backends that require the plugin to implement this trait ensure that when calling the render_buffer method of the AudioRenderer trait

the number of inputs channels (buffer.number_of_input_channels()) is smaller than or equal to Self::max_number_of_audio_inputs() and

the number of outputs (buffer.number_of_output_channels()) is smaller than or equal to Self::max_number_of_audio_outputs().

Remark

This trait can be more conveniently implemented by implementing the Meta trait.

Required methods

fn max_number_of_audio_inputs(&self) -> usize

The maximum number of audio inputs supported. This method should return the same value every time it is called.

fn max_number_of_audio_outputs(&self) -> usize

The maximum number of audio outputs supported. This method should return the same value every time it is called.

Loading content...