Trait rsynth::CommonMidiPortMeta

pub trait CommonMidiPortMeta: MidiHandlerMeta {
    fn midi_input_name(&self, index: usize) -> String { ... }

    fn midi_output_name(&self, index: usize) -> String { ... }
}

Provides some meta-data of the midi-ports used by the plugin or application to the host. This trait can be more conveniently implemented by implementing the Meta trait.

Provided methods

fn midi_input_name(&self, index: usize) -> String

The name of the midi input with the given index. You can assume that index is strictly smaller than Self::max_number_of_midi_inputs().

Note

When using the Jack backend, this function should not return an empty string.

fn midi_output_name(&self, index: usize) -> String

The name of the midi output with the given index. You can assume that index is strictly smaller than Self::max_number_of_midi_outputs()

Note

When using the Jack backend, this function should not return an empty string.

Loading content...