Trait rsynth::ContextualAudioRenderer

pub trait ContextualAudioRenderer<F, Context>: AudioHandler {
    fn render_buffer(
        &mut self, 
        inputs: &[&[F]], 
        outputs: &mut [&mut [F]], 
        context: &mut Context
    );
}

Defines how audio is rendered, similar to the AudioRenderer trait. The extra parameter context can be used by the backend to provide extra information.

See the documentation of [AudioRenderer] for more information.

Required methods

fn render_buffer(
&mut self,
inputs: &[&[F]],
outputs: &mut [&mut [F]],
context: &mut Context
)

This method called repeatedly for subsequent buffers.

It is similar to the [render_buffer] from the [AudioRenderer] trait, see its documentation for more information.

Loading content...