Trait rsynth::MidiHandlerMeta

pub trait MidiHandlerMeta {
    fn max_number_of_midi_inputs(&self) -> usize;

    fn max_number_of_midi_outputs(&self) -> usize;
}

Define the maximum number of midi inputs and the maximum number of midi outputs. This trait can be more conveniently implemented by implementing the Meta trait.

Required methods

fn max_number_of_midi_inputs(&self) -> usize

The maximum number of midi inputs supported. This method should return the same value for subsequent calls.

fn max_number_of_midi_outputs(&self) -> usize

The maximum number of midi outputs supported. This method should return the same value for subsequent calls.

Loading content...