Trait rsynth::Plugin

pub trait Plugin<C> {
    const NAME: &'static str;
    const MAX_NUMBER_OF_AUDIO_INPUTS: usize;
    const MAX_NUMBER_OF_AUDIO_OUTPUTS: usize;

    fn audio_input_name(index: usize) -> String;

    fn audio_output_name(index: usize) -> String;

    fn set_sample_rate(&mut self, sample_rate: f64);

    fn render_buffer<F>(
        &mut self, 
        inputs: &[&[F]], 
        outputs: &mut [&mut [F]], 
        context: &mut C
    )
    where
        F: Float + AsPrim;
}

The trait that all plugins need to implement. The parameter C corresponds to the context of the plugin.

Associated Constants

const NAME: &'static str

The name of the plugin.

const MAX_NUMBER_OF_AUDIO_INPUTS: usize

The maximum number of audio inputs.

const MAX_NUMBER_OF_AUDIO_OUTPUTS: usize

The maximum number of audio outputs.

Loading content...
Required methods

fn audio_input_name(index: usize) -> String

The name of the audio input with the given index. Note: you may not provide an empty string to the Jack backend.

fn audio_output_name(index: usize) -> String

The name of the audio output with the given index. Note: you may not provide an empty string to the Jack backend.

fn set_sample_rate(&mut self, sample_rate: f64)

Called when the sample-rate changes. The backend should ensure that this function is called before any other.

fn render_buffer<F>(
&mut self,
inputs: &[&[F]],
outputs: &mut [&mut [F]],
context: &mut C
) where
F: Float + AsPrim,

This function is the core of the plugin. It is called repeatedly for subsequent buffers. The length of inputs is guaranteed to be smaller than or equal to Self::MAX_NUMBER_OF_AUDIO_INPUTS. The length of outputs is guaranteed to be smaller than or equal to Self::MAX_NUMBER_OF_AUDIO_OUTPUTS. The lengths of all elements of inputs and the lengths of all elements of outputs are all guaranteed to equal to each other. This shared length can however be different for subsequent calls to render_buffer.

Loading content...