Struct rsynth::utilities::polyphony::simple_event_dispatching::SimpleEventDispatcher

pub struct SimpleEventDispatcher<Classifier, V> { /* fields omitted */ }

A simple event dispatcher.

The type parameter Classifier refers to the classifier that is used to classify events. In order to use this SimpleEventDispatcher, the concrete type used for Classifier should implement the EventDispatchClassifier trait.

The type parameter V refers to the voice.

Methods

impl<Classifier, V> SimpleEventDispatcher<Classifier, V>

pub fn new(classifier: Classifier) -> Self