Crate rsynth

Rsynth

A crate for developing audio plugins and applications in Rust, with a focus on software synthesis. Rsynth is well suited as a bootstrap for common audio plugin generators. It handles voices, voice-stealing, polyphony, etc. so the programmer's main focus can be DSP.

Back-ends

rsynth currently supports the following back-ends:

combined combine different back-ends for audio input, audio output, midi input and midi output, mostly for offline rendering and testing (behind various features)

jack (behind the backend-jack feature)

vst (behind the backend-vst)

See the documentation of each back-end for more information.

Rendering audio

Audio can be rendered with the [ContextualAudioRenderer] trait that is generic over the floating point type (f32 or f64). The parameter context is used by the host or environment to pass extra data.

The plugin or application can internally also use the AudioRenderer trait, which is similar to the [ContextualAudioRenderer] trait, but does not have a context parameter.

Meta-data

There are a number of traits to define some meta-data. Every plugin should implement these, but it can be tedious, so you can implement these traits in a more straightforward way by implementing the Meta trait. However, you can also implement these trait "by hand":

CommonPluginMeta

Name of the plugin etc

AudioHandlerMeta

Number of audio ports

MidiHandlerMeta

Number of midi ports

CommonAudioPortMeta

Names of the audio in and out ports

CommonPluginMeta

Name of the plugin or application

Handling events

Plugins and applications can also implement ContextualEventHandler and EventHandler for each event type that they support. Currently supported events are:

RawMidiEvent

SysExEvent

Timed<T>

Indexed<T>

Utilities

Utilities are are types that you can include to perform several common tasks for the plugin or application:

polyphony: managing of different voices

timesplitting: split the audio buffer at the events

Some audio concepts

A sample is a single number representing the air pressure at a given time. It is usually represented by an f32, f64, i16 or i32 number, but other types are possible as well.

A channel usually corresponds with a speaker or a number of speakers. E.g. in a stereo setup, there is a "left" channel and a "right" channel.

A frame consists of the samples for all the channels at a given time.

A buffer consists of subsequent samples for a given channel and corresponds to a certain time period. (Non-standard terminology.)

A chunk consists of the buffers for all channels for a given time period. (Non-standard terminology.)

                         ┌ chunk     ┌ frame
             ┌ sample    ↓           ↓
             │      ┌─────────┐     ┌─┐
          ┌──↓──────┼─────────┼─────┼─┼───────────────────┐
 channel →│• • • • •│• • • • •│• • •│•│• • • • • • • • • •│
          └─────────┼─────────┼─────┼─┼───────────────────┘
           • • • • •│• • • • •│• • •│•│• • • • • • • • • •
                    │         │     │ │   ┌───────┐
           • • • • •│• • • • •│• • •│•│• •│• • • •│• • • •
                    └─────────┘     └─┘   └───────┘
                                            ↑
                                            └ buffer

Modules

backend
Backends.

buffer

envelope

event
Event handling

meta
Example

middleware
Middleware like support for polyphony, ...

test_utilities
Utilities for testing.

utilities

Macros

audio_chunk

vst_init
A wrapper around the plugin_main! macro from the vst crate. You call this with one parameter, which is the function declaration of a function that creates your plugin. This function may also do some setup (e.g. initialize logging).

Traits

AudioHandler
Define how sample-rate changes are handled.

AudioHandlerMeta
Define the maximum number of audioinputs and the maximum number of audio outputs. This trait can be more conveniently implemented by implementing the Meta trait.

AudioRenderer
Defines how audio is rendered.

CommonAudioPortMeta
Provides some meta-data of the audio-ports used by the plugin or application to the host. This trait can be more conveniently implemented by implementing the Meta trait.

CommonMidiPortMeta
Provides some meta-data of the midi-ports used by the plugin or application to the host. This trait can be more conveniently implemented by implementing the Meta trait.

CommonPluginMeta
Provides common meta-data of the plugin or application to the host. This trait is common for all backends that need this info. This trait can be more conveniently implemented by implementing the Meta trait.

ContextualAudioRenderer
Defines how audio is rendered, similar to the AudioRenderer trait. The extra parameter context can be used by the backend to provide extra information.

MidiHandlerMeta
Define the maximum number of midi inputs and the maximum number of midi outputs. This trait can be more conveniently implemented by implementing the Meta trait.