Trait rsynth::backend::HostInterface

pub trait HostInterface {
    fn output_initialized(&self) -> bool;
}

Defines an interface for communicating with the host or server of the backend, e.g. the VST host when using VST or the Jack server when using Jack.

Required Methods

fn output_initialized(&self) -> bool

Return whether the output buffers are zero-initialized. Returns false when in doubt.

Example

The following example illustrates how output_initialized() can be used in combination with rsynth::utilities::zero_init to initialize the output buffers to zero in an implementation of the ContextualAudioRenderer trait.

use rsynth::ContextualAudioRenderer;
use rsynth::backend::HostInterface;
use rsynth::utilities::initialize_to_zero;
impl<H> ContextualAudioRenderer<f32, H> for MyPlugin
where H: HostInterface
{
    fn render_buffer(
        &mut self,
        inputs: &[&[f32]],
        outputs: &mut [&mut [f32]],
        context: &mut H)
    {
        if ! context.output_initialized() {
            initialize_to_zero(outputs);
            // The rest of the audio rendering.
        }
    }
}