Module rsynth::backend::vst_backend

Wrapper for the VST backend.

Support is only enabled if you compile with the "vst-backend" feature, see the cargo reference for more information on setting cargo features.

For an example, see vst_synth.rs in the examples folder. examples/test_synth.rs contains the code that is shared for all backends and examples/vst_synth.rs contains the jack-specific code.

Usage

See also the documentation of the vst_init macro.

Structs

VstPluginWrapper
A struct used internally by the vst_init macro. Normally, plugin's do not need to use this.

Traits

VstPlugin
A VST plugin should implement this trait in addition to the Plugin trait.