Trait rsynth::envelope::Envelope

pub trait Envelope<'a, T> {
    type Iter: Iterator<Item = EnvelopeIteratorItem<T>>;
    type EventType;
    fn iter(&'a self) -> Self::Iter;

    fn insert_event(&mut self, event: Self::EventType);

    fn forget_past(&mut self, number_of_frames_to_forget: u32);
}

Defines the behaviour of an envelope. An envelope allows to get an iterator. The returned iterator allows to iterator over the frames, starting from the current position, and for each frame, returns the envelope value at that frame.

Associated Types

type Iter: Iterator<Item = EnvelopeIteratorItem<T>>

The type of the iterator.

type EventType

Loading content...
Required methods

fn iter(&'a self) -> Self::Iter

Get the iterator.

fn insert_event(&mut self, event: Self::EventType)

fn forget_past(&mut self, number_of_frames_to_forget: u32)

Loading content...