Struct rsynth::event::event_queue::EventQueue

pub struct EventQueue<T> { /* fields omitted */ }

Methods

impl<T> EventQueue<T>

pub fn new(capacity: usize) -> Self

Panics

Panics if capacity == 0.

pub fn queue_event<H>(
&mut self,
new_event: Timed<T>,
collision_decider: H
) -> Option<Timed<T>> where
H: HandleEventCollision<T>,

Queue a new event. When the buffer is full, an element may be removed from the queue to make some room. This element is returned.

pub fn forget_before(&mut self, threshold: u32) where
T: Copy,

Remove all events before, but not on, this threshold.

Note about usage in real-time context

If T implements drop, the elements that are removed are dropped. This may cause memory de-allocation, which you want to avoid in the real-time part of your library.

pub fn clear(&mut self)

Remove all events from the queue.

Note about usage in real-time context

If T implements drop, the elements that are removed are dropped. This may cause memory de-allocation, which you want to avoid in the real-time part of your library.

pub fn shift_time(&mut self, new_zero_time: u32)

Shift time forward by new_zero_time frames.

Panics

Panics in debug mode when at least one event has a time_in_frames that is < new_zero_time.

pub fn get_last_before(&self, time: u32) -> Option<&Timed<T>>

pub fn first(&self) -> Option<&Timed<T>>

pub fn split<'in_storage, 'out_storage, 'in_channels, 's, 'chunk, S, R, C>(
&mut self,
input_storage: &'in_storage mut VecStorage<&'static [S]>,
output_storage: &'out_storage mut VecStorage<&'static mut [S]>,
buffer: &mut AudioBufferInOut<'in_channels, '_, '_, '_, S>,
renderer: &mut R,
context: &mut C
) where
S: Copy + 'static,
R: ContextualAudioRenderer<S, C> + EventHandler<T>,
T: Debug,