Struct rsynth::event::event_queue::EventQueue

pub struct EventQueue<T> { /* fields omitted */ }

Methods

impl<T> EventQueue<T>

pub fn new(capacity: usize) -> Self

pub fn len(&self) -> usize

pub fn queue_event<H>(
&mut self,
new_event: Timed<T>,
collision_decider: H
) -> Option<Timed<T>> where
H: HandleEventCollision<T>,

Queue a new event. When the buffer is full, an element may be removed from the queue to make some room. This element is returned.

pub fn forget_before(&mut self, threshold: u32) where
T: Copy,

Remove all events before, but not on, this threshold.

Allocation

If T implements drop, the elements that are removed are dropped. This may cause memory de-allocation, which you want to avoid in the real-time part of your library.

pub fn clear(&mut self)

Remove all events from the queue.

Allocation

If T implements drop, the elements that are removed are dropped. This may cause memory de-allocation, which you want to avoid in the real-time part of your library.

pub fn shift_time(&mut self, new_zero_time: u32)

Shift time forward by new_zero_time frames.

Panics

Panics in debug mode when at least one event has a time_in_frames that is < new_zero_time.

pub fn get_last_before(&self, time: u32) -> Option<&Timed<T>>

pub fn first(&self) -> Option<&Timed<T>>

Important traits for Iter<'a, T>
impl<'a, T> Iterator for Iter<'a, T> type Item = &'a T;

pub fn iter<'a>(&'a self) -> Iter<'a, Timed<T>>