Trait rsynth::event::EventHandler

pub trait EventHandler<E> {
    fn handle_event(&mut self, event: E);
}

The trait that plugins should implement in order to handle the given type of events.

The type parameter E corresponds to the type of the event.

Required methods

fn handle_event(&mut self, event: E)

Loading content...