Trait rsynth::CommonPluginMeta

pub trait CommonPluginMeta {
    fn name(&self) -> &str;
}

Provides common meta-data of the plugin or application to the host. This trait is common for all backends that need this info. This trait can be more conveniently implemented by implementing the Meta trait.

Required methods

fn name(&self) -> &str

The name of the plugin or application.

Loading content...