Trait rsynth::CommonPluginMeta

pub trait CommonPluginMeta {
    fn name(&self) -> &str { ... }

    fn plugin_name<W: Write>(&self, buffer: &mut W) -> Result<(), Error> { ... }
}

Provides common meta-data of the plugin or application to the host. This trait is common for all backends that need this info. This trait can be more conveniently implemented by implementing the Meta trait.

Provided methods

fn name(&self) -> &str

The name of the plugin or application.

#[deprecated(since = "0.1.2", note = "Use or implement [plugin_name] instead.")]

fn plugin_name<W: Write>(&self, buffer: &mut W) -> Result<(), Error>

The name of the plugin or application.

Compatibility note

The default implementation of this method will likely be removed in a future release.

Loading content...