Trait rsynth::ContextualAudioRenderer

pub trait ContextualAudioRenderer<S, Context> where
    S: Copy, {
    fn render_buffer(
        &mut self, 
        buffer: &mut AudioBufferInOut<'_, '_, '_, '_, S>, 
        context: &mut Context
    );
}

Defines how audio is rendered, similar to the AudioRenderer trait. The extra parameter context can be used by the backend to provide extra information.

The type parameter S refers to the data type of a sample. It is typically f32 or f64.

Required methods

fn render_buffer(
&mut self,
buffer: &mut AudioBufferInOut<'_, '_, '_, '_, S>,
context: &mut Context
)

This method called repeatedly for subsequent buffers.

It is similar to the render_buffer from the AudioRenderer trait, see its documentation for more information.

Loading content...